skip unneeded proc unit for postpipeline
1. some post proc units like digital zoom is not needed for every frame,
so we could skip them if possible.
2. fix PostProcessPipeline::prepare bugs
It adds a proc unit to middle level and last level at the same time,
and one unit could only belong to one level. If the whole post
pipe line lenth is 1, then the unit could belong to first or last level,
and now we set it to first level besides adding frameworker notify callback
to it.

Change-Id: I779880525c70da4bd57e40c92ed0a8a1c1cf7fee
Signed-off-by: ZhongYichong <zyc@rock-chips.com>