set pipeline depth to 6 for 3126c
because the performance of RGA of rk3126c is not satisfied for big
frame size like 1600x1200, and this will cause frame dropped issue,
so we increase more cache buffers to avoid driver dropping the frame
caused by no queued buffer.

Should NOTICE that CIF uses the CMA memory, so the default CMA memory
size should be adjusted to satisfy the pipeline depth. For 1600x1200,
we need more than 21M CMA memory.

Change-Id: I958ac3d3322302f1cbf8d66acb632b5316608a0b
Signed-off-by: ZhongYichong <zyc@rock-chips.com>