fix wrong file copy path for under android9
HAL1 and HAL3 in android version lower than 9.0 can exist at the
same time, and we use folder camera/ to place HAL1 sources while
camera_v3/ to place HAL3 sources. For android version equal and
upper than 9.0, HAL1 is not allowed and HAL3 is mandatory, so we
use camera/ to place HAL3 sources. So we should distinguish the
file copy path in camera_etc.mk according to android version.

Change-Id: I5e0ca0da561ac7a597f819ab9deddf975795dfd8
Signed-off-by: ZhongYiChong <zyc@rock-chips.com>